# Optimize lines_lf