--- a/src/lines_lf.rs
+++ b/src/lines_lf.rs
@@ -47,26 +47,12 @@ pub fn to_byte_idx(text: &str, line_idx: usize) -> usize {
 }
 
 //-------------------------------------------------------------
-const LF: u8 = b'\n';
 
 #[inline(always)]
 fn to_byte_idx_impl<T: ByteChunk>(text: &[u8], line_idx: usize) -> usize {
     let mut byte_count = 0;
     let mut lf_count = 0;
 
-    if text.len() <= T::SIZE {
-        // Bypass the more complex routine for short strings, where the
-        // complexity hurts performance.
-        for (i, byte) in text.iter().enumerate() {
-            if lf_count == line_idx {
-                return i;
-            }
-            if *byte == LF {
-                lf_count += 1;
-            }
-        }
-        return text.len();
-    }
     // Get `middle` so we can do more efficient chunk-based counting.
     // We can't use this to get `end`, however, because the start index of
     // `end` actually depends on the accumulating line counts during the
@@ -78,7 +64,7 @@ fn to_byte_idx_impl<T: ByteChunk>(text: &[u8], line_idx: usize) -> usize {
         if lf_count == line_idx {
             return byte_count;
         }
-        if *byte == LF {
+        if *byte == 0x0A {
             lf_count += 1;
         }
         byte_count += 1;
@@ -87,10 +73,10 @@ fn to_byte_idx_impl<T: ByteChunk>(text: &[u8], line_idx: usize) -> usize {
     // Process the chunks 4 at a time
     let mut chunk_count = 0;
     for chunks in middle.chunks_exact(4) {
-        let val1 = chunks[0].cmp_eq_byte(LF);
-        let val2 = chunks[1].cmp_eq_byte(LF);
-        let val3 = chunks[2].cmp_eq_byte(LF);
-        let val4 = chunks[3].cmp_eq_byte(LF);
+        let val1 = chunks[0].cmp_eq_byte(0x0A);
+        let val2 = chunks[1].cmp_eq_byte(0x0A);
+        let val3 = chunks[2].cmp_eq_byte(0x0A);
+        let val4 = chunks[3].cmp_eq_byte(0x0A);
         let new_lf_count = lf_count + val1.add(val2).add(val3.add(val4)).sum_bytes();
         if new_lf_count >= line_idx {
             break;
@@ -102,7 +88,7 @@ fn to_byte_idx_impl<T: ByteChunk>(text: &[u8], line_idx: usize) -> usize {
 
     // Process the rest of the chunks
     for chunk in middle[chunk_count..].iter() {
-        let new_lf_count = lf_count + chunk.cmp_eq_byte(LF).sum_bytes();
+        let new_lf_count = lf_count + chunk.cmp_eq_byte(0x0A).sum_bytes();
         if new_lf_count >= line_idx {
             break;
         }
@@ -115,9 +101,7 @@ fn to_byte_idx_impl<T: ByteChunk>(text: &[u8], line_idx: usize) -> usize {
         if lf_count == line_idx {
             break;
         }
-        if *byte == LF {
-            lf_count += 1;
-        }
+        lf_count += (*byte == 0x0A) as usize;
         byte_count += 1;
     }
 
@@ -133,7 +117,7 @@ fn count_breaks_impl<T: ByteChunk>(text: &[u8]) -> usize {
     if text.len() < T::SIZE {
         // Bypass the more complex routine for short strings, where the
         // complexity hurts performance.
-        count_lf(text)
+        text.iter().map(|byte| (*byte == 0x0A) as usize).sum()
     } else {
         // Get `middle` so we can do more efficient chunk-based counting.
         let (start, middle, end) = unsafe { text.align_to::<T>() };
@@ -141,31 +125,34 @@ fn count_breaks_impl<T: ByteChunk>(text: &[u8]) -> usize {
         let mut count = 0;
 
         // Take care of unaligned bytes at the beginning.
-        count += count_lf(start);
+        count += start
+            .iter()
+            .map(|byte| (*byte == 0x0A) as usize)
+            .sum::<usize>();
 
         // Take care of the middle bytes in big chunks. Loop unrolled.
         for chunks in middle.chunks_exact(4) {
-            let val1 = chunks[0].cmp_eq_byte(LF);
-            let val2 = chunks[1].cmp_eq_byte(LF);
-            let val3 = chunks[2].cmp_eq_byte(LF);
-            let val4 = chunks[3].cmp_eq_byte(LF);
+            let val1 = chunks[0].cmp_eq_byte(0x0A);
+            let val2 = chunks[1].cmp_eq_byte(0x0A);
+            let val3 = chunks[2].cmp_eq_byte(0x0A);
+            let val4 = chunks[3].cmp_eq_byte(0x0A);
             count += val1.add(val2).add(val3.add(val4)).sum_bytes();
         }
 
         // Chunk remainder
         let mut acc = T::zero();
         for chunk in middle.chunks_exact(4).remainder() {
-            acc = acc.add(chunk.cmp_eq_byte(LF));
+            acc = acc.add(chunk.cmp_eq_byte(0x0A));
         }
         count += acc.sum_bytes();
 
         // Take care of unaligned bytes at the end.
-        count + count_lf(end)
+        count
+            + end
+                .iter()
+                .map(|byte| (*byte == 0x0A) as usize)
+                .sum::<usize>()
     }
-}
-
-fn count_lf(text: &[u8]) -> usize {
-    text.iter().filter(|&b| *b == LF).count()
 }
 
 //=============================================================